Fix : 댓글 부분 컴포넌트화하여 댓글이 해당 포스트에만 업로드되도록 수정
