+++ src/pages/kris/Main/MainKris.scss
@@ -107,128 +107,6 @@ input:focus {
     background-color: #f1f1f1;
   }
 
-  .feed-article {
-    display: flex;
-    flex-direction: column;
-    width: 500px;
-    min-width: 500px;
-    margin-bottom: 20px;
-    border: 1px solid lightgrey;
-    background-color: white;
-  }
-
-  .feed-info {
-    display: flex;
-    align-items: center;
-
-    padding: 15px;
-  }
-
-  .feed-info-profile {
-    width: 32px;
-    height: 32px;
-    margin: 0 10px 0 0;
-    overflow: hidden;
-    border-radius: 70%;
-    object-fit: cover;
-  }
-
-  .feed-info-profile img {
-    width: 100%;
-  }
-
-  .feed-info-author a {
-    color: black;
-    font-family: sans-serif;
-    font-weight: 600;
-  }
-
-  .feed-image img {
-    width: 100%;
-  }
-
-  .feed-interaction {
-    display: flex;
-    margin-top: 10px;
-    padding-left: 10px;
-  }
-
-  .interaction-icon {
-    height: 25px;
-    margin-right: 15px;
-  }
-
-  .feed-text {
-    margin: 10px 0 5px 0;
-    padding: 0 10px;
-    font-family: sans-serif;
-  }
-
-  .feed-text-author {
-    margin-right: 10px;
-    font-weight: 600;
-  }
-
-  .comment-posted {
-    display: flex;
-    align-items: first baseline;
-    justify-content: space-between;
-    margin-top: 10px;
-
-    padding: 0 10px;
-    font-family: sans-serif;
-  }
-
-  .comment-posted-author {
-    margin-right: 10px;
-    font-weight: 600;
-  }
-
-  .comment-posted-text {
-    width: 320px;
-    margin-right: 20px;
-  }
-
-  .comment-posted i {
-    margin-right: 5px;
-    color: #757575;
-    font-size: 12px;
-  }
-
-  .comment-posted button {
-    width: 30px;
-    border: transparent;
-    background-color: transparent;
-    color: #6c6c6c;
-    font-size: 10px;
-    font-weight: 600;
-  }
-
-  .feed-add-comment {
-    display: flex;
-    justify-content: space-between;
-    height: 40px;
-    margin-top: 10px;
-
-    padding: 0 10px;
-  }
-
-  .feed-add-comment input {
-    width: 410px;
-    border: transparent;
-    background-color: transparent;
-    font-size: 15px;
-  }
-
-  .feed-add-comment button {
-    width: 40px;
-    border: transparent;
-    background-color: transparent;
-    color: #0995f6;
-    font-size: 15px;
-    font-weight: 600;
-  }
-
   .main-right {
     width: 320px;
   }
